fix(auth): do not perform auth on auth files
Auth files should be loaded with having to pass auth checks
in the trust chain, so that arbitrary certs can be loaded
for PK/KEK/db/dbx, therefore load auths in SETUP_MODE, then
update the platform mode to USER if there has been a PK set.

The PK could be set by a backend or by finding an auth file.

Signed-off-by: Bobby Eshleman <bobby.eshleman@gmail.com>