feat(logging): make log level read from config var
This commit adds support for changing the log level by defining it with
the config variable "loglevel" in file
"/etc/uefistored/uefistored.conf".

The accepted values for the config variable "loglevel"
are "ERROR", "WARNING", "DEBUG", or "INFO".

This commit uses the config parsing routine from glib2 for
parsing the configuration file.

If parsing fails or fails to find the loglevel defined in the config
file, then the loglevel defaults to INFO.

Adds glib2 as a dependency.

Signed-off-by: Bobby Eshleman <bobby.eshleman@gmail.com>