fix(secureboot-certs): remove python-requests dependency
The python-requests package is not installed by default in dom0.
Instead of requiring it as a dependency, this commit uses of Python's
standard urllib2, which should not add any further dependencies.

urllib2 is a little less ergonomic as the downloads in secureboot-certs
fails if certain headers are not defined (unlike the requests library,
which sets them itself).  The rationale behind those headers is
commented inline.

Signed-off-by: Bobby Eshleman <bobby.eshleman@gmail.com>