fix(auth): don't append zero data when sigs filter to zero
`FilterSignatureList()` is meant to prevent the appending of duplicate
certificates to variable which already contains those certificates. It
returns the list of certs that are actually new and the data size of the
blob containing them. If there are no new certs, it'll be a data size of
zero.  Appending zero data to a variable is not necessary, so simply
exit early in this case.

Signed-off-by: Bobby Eshleman <bobby.eshleman@gmail.com>