feat(secureboot-certs): accept auth, PEM X509, or DER X509 format files
This commit adds support for secureboot-certs to accept any of the
following formats as input files:
    - EFI auth
    - DER-encoded X509
    - PEM-encoded X509

secureboot-certs internally converts all of these file formats into an
auth file.  Besides making life easier for users, this also allows users
to put multiple certificates together into a single auth file and pass
in that auth file to secureboot-certs.

This commit does NOT add support for passing in multiple PEM/DER encoded
X509 certificates for a single UEFI variable. If a user wishes to do
this, they must use a tool like `create-auth` or `efitools` to bundle
the X509 certificates into a single EFI auth file and pass that to
secureboot-certs.

Additionally, setting a custom PK cert also requires that the user pass
the private half of the certificate to secureboot-certs (in the argument
--pk-priv).  This is because the PK must be self-signed according to the
UEFI spec and this quality is enforced by uefistored.  Passing a PK.auth
to secureboot-certs does not require --pk-priv because the auth file has
already been created at this point, and so it is assumed it was
self-signed, which is safe assumption given that both efitools and
create-auth both do it this way.

Signed-off-by: Bobby Eshleman <bobby.eshleman@gmail.com>