fix: allow no db and no dbx, but still prevent bad configs
Allowing a PK to be installed with no KEK creates systems in which users
can not install a db/dbx. This matters for Windows especially, because
it tries to install a dbx when updating. This was fixed but loading the
certs in order: db -> dbx -> KEK -> PK, and if any preceding certs
failed to load in the chain, then no following certs would be loaded.
Unfortunately, this also placed an implicit requirement that systems
have both a db and dbx, which should not be required... A user may
simply want the MSFT KEK and a default PK on the system, and wants to
allow Windows to install the dbx later at run time.

Signed-off-by: Bobby Eshleman <bobby.eshleman@gmail.com>