logging: demote or remove superfluous error messages
This commit either reduces overzealous ERROR() messages to WARNING()
messages. If the message occurs during normal operation of a VM (such as
uefistored receiving an IO event that it doesn't know how to handle),
then the message is removed altogether.

Finally, a newline char was added to a message that was missing one.

Signed-off-by: Bobby Eshleman <bobby.eshleman@gmail.com>